requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
bcrypt==4.1.2
requests==2.31.0
python-dotenv==1.0.1
pytest==8.0.0